some cleanup & file renaming
