--- x/gmp/types/gmp_message_types.go
@@ -1,12 +1,0 @@
-package types
-
-const (
-	// TypeUnrecognized means coin type is unrecognized
-	TypeUnrecognized = iota
-	// TypeGeneralMessage is a pure message
-	TypeGeneralMessage
-	// TypeGeneralMessageWithToken is a general message with token
-	TypeGeneralMessageWithToken
-	// TypeSendToken is a direct token transfer
-	TypeSendToken
-)
